benchmark: split script into files and add to entrypoints
closes #553
