+++ reana/reana_benchmark/__init__.py
@@ -1,0 +1,7 @@
+# This file is part of REANA.
+# Copyright (C) 2021 CERN.
+#
+# REANA is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""reana-benchmark - benchmark script for REANA platform."""
