+++ reana/reana_benchmark/utils.py
@@ -1,0 +1,23 @@
+# This file is part of REANA.
+# Copyright (C) 2021 CERN.
+#
+# REANA is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""Contains common functions shared by reana-benchmark modules."""
+
+import logging
+from datetime import datetime
+
+from reana.reana_benchmark.config import DATETIME_FORMAT
+
+logger = logging.getLogger("reana-benchmark")
+logger.setLevel(logging.INFO)
+
+
+def get_utc_now_timestamp() -> str:  # noqa: D103
+    return datetime.utcnow().strftime(DATETIME_FORMAT)
+
+
+def build_extended_workflow_name(workflow: str, run_number: int) -> str:  # noqa: D103
+    return f"{workflow}-{run_number}"
